backend/requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langgraph>=0.2.0
chromadb>=0.4.0
sqlitevec
python-dotenv>=1.0.0
apscheduler>=3.10.0
fastapi>=0.100.0
uvicorn>=0.25.0
pydantic>=2.0.0
tiktoken>=0.5.0
playwright>=1.40.0
httpx>=0.25.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pyyaml>=6.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
dashscope>=1.20.0
